sync

+++ module/std/private/std_byte.c
@@ -322,7 +322,7 @@ uint32_t std_min_u32 ( uint32_t a, uint32_t b ) {
     return a < b ? a : b;
 }
 
-uint32_t std_min_u64 ( uint32_t a, uint32_t b ) {
+uint64_t std_min_u64 ( uint64_t a, uint64_t b ) {
     return a < b ? a : b;
 }
