+++ module/std/public/std_byte.h
@@ -134,7 +134,7 @@ size_t      std_min     ( size_t a, size_t b );
 int32_t     std_min_i32 ( int32_t a, int32_t b );
 int64_t     std_min_i64 ( int64_t a, int64_t b );
 uint32_t    std_min_u32 ( uint32_t a, uint32_t b );
-uint32_t    std_min_u64 ( uint32_t a, uint32_t b );
+uint64_t    std_min_u64 ( uint64_t a, uint64_t b );
 float       std_min_f32 ( float a, float b );
 double      std_min_f64 ( double a, double b );
